split uinodes into specific files
