+++ ShaderTool/Shadertool.vcxproj
@@ -98,9 +98,9 @@
   <ItemGroup>
     <ClInclude Include="src\AssetManager.h" />
     <ClInclude Include="src\Defines.h" />
-    <ClInclude Include="src\Editor\Edge.h" />
-    <ClInclude Include="src\Editor\Graph.h" />
-    <ClInclude Include="src\Editor\IdMap.h" />
+    <ClInclude Include="src\Editor\Graph\Edge.h" />
+    <ClInclude Include="src\Editor\Graph\Graph.h" />
+    <ClInclude Include="src\Editor\Graph\IdMap.h" />
     <ClInclude Include="src\Editor\ImGui\ImGuiFileBrowser\Dirent\dirent.h" />
     <ClInclude Include="src\Editor\ImGui\ImGuiFileBrowser\ImGuiFileBrowser.h" />
     <ClInclude Include="src\Editor\ImGui\ImGuiFileDialog\ImGuiFileDialog.h" />
@@ -117,12 +117,20 @@
     <ClInclude Include="src\Editor\ImGui\imstb_rectpack.h" />
     <ClInclude Include="src\Editor\ImGui\imstb_textedit.h" />
     <ClInclude Include="src\Editor\ImGui\imstb_truetype.h" />
-    <ClInclude Include="src\Editor\nfd\common.h" />
-    <ClInclude Include="src\Editor\nfd\nfd.h" />
-    <ClInclude Include="src\Editor\nfd\nfd_common.h" />
-    <ClInclude Include="src\Editor\nfd\simple_exec.h" />
-    <ClInclude Include="src\Editor\Node.h" />
-    <ClInclude Include="src\Editor\UiNode.h" />
+    <ClInclude Include="src\Editor\NFD\common.h" />
+    <ClInclude Include="src\Editor\NFD\nfd.h" />
+    <ClInclude Include="src\Editor\NFD\nfd_common.h" />
+    <ClInclude Include="src\Editor\NFD\simple_exec.h" />
+    <ClInclude Include="src\Editor\Graph\Node.h" />
+    <ClInclude Include="src\Editor\UiNode\AddNode.h" />
+    <ClInclude Include="src\Editor\UiNode\DrawNode.h" />
+    <ClInclude Include="src\Editor\UiNode\MultiplyNode.h" />
+    <ClInclude Include="src\Editor\UiNode\PrimitiveNode.h" />
+    <ClInclude Include="src\Editor\UiNode\RenderTargetNode.h" />
+    <ClInclude Include="src\Editor\UiNode\ShaderNode.h" />
+    <ClInclude Include="src\Editor\UiNode\SineNode.h" />
+    <ClInclude Include="src\Editor\UiNode\TimeNode.h" />
+    <ClInclude Include="src\Editor\UiNode\UiNode.h" />
     <ClInclude Include="src\Entity.h" />
     <ClInclude Include="src\GameTimer.h" />
     <ClInclude Include="src\GeometryGenerator.h" />
@@ -174,10 +182,10 @@
     <ClCompile Include="src\Editor\ImGui\imnodes.cpp">
       <PrecompiledHeader>NotUsing</PrecompiledHeader>
     </ClCompile>
-    <ClCompile Include="src\Editor\nfd\nfd_common.c">
+    <ClCompile Include="src\Editor\NFD\nfd_common.c">
       <PrecompiledHeader>NotUsing</PrecompiledHeader>
     </ClCompile>
-    <ClCompile Include="src\Editor\nfd\nfd_win.cpp">
+    <ClCompile Include="src\Editor\NFD\nfd_win.cpp">
       <PrecompiledHeader>NotUsing</PrecompiledHeader>
     </ClCompile>
     <ClCompile Include="src\GameTimer.cpp" />
