+++ ShaderTool/src/Rendering/ShaderManager.h
@@ -16,8 +16,8 @@ public:
 		return instance;
 	}
 
-	bool LoadBinaryShader(const std::string& name);
-	bool LoadRawShader(const std::string& filename, const std::string& entryPoint, const std::string& target);
+	int LoadBinaryShader(const std::string& name);
+	int LoadRawShader(const std::string& filename, const std::string& entryPoint, const std::string& target);
 	Shader* ShaderManager::GetShader(size_t index);
 	Shader* GetShader(const std::string& name);
